@@ -4,6 +4,8 @@
   "scripts": {
     "dev": "next dev -p 3001",
     "build": "next build",
+    "build:debug": "next build --debug",
+    "build:profile": "next build --profile",
     "start": "next start",
     "start:3001": "next start -p 3001",
     "test": "jest",
@@ -15,19 +17,19 @@
   "dependencies": {
     "@apollo/client": "^3.6.2",
     "@elastic/elasticsearch": "^7.17.0",
-    "@emotion/cache": "^11.11.0",
-    "@emotion/react": "^11.11.3",
+    "@emotion/cache": "^11.13.1",
+    "@emotion/react": "^11.13.3",
     "@emotion/server": "^11.11.0",
-    "@emotion/styled": "^11.11.0",
-    "@mui/base": "^5.0.0-beta.33",
-    "@mui/icons-material": "^5.15.9",
-    "@mui/lab": "^5.0.0-alpha.165",
-    "@mui/material": "^5.15.9",
-    "@mui/material-nextjs": "^5.15.9",
-    "@mui/system": "^5.15.11",
+    "@emotion/styled": "^11.13.0",
+    "@mui/base": "^5.0.0-beta.61",
+    "@mui/icons-material": "^6.1.6",
+    "@mui/lab": "^6.0.0-beta.14",
+    "@mui/material": "^6.1.6",
+    "@mui/material-nextjs": "^6.1.6",
+    "@mui/system": "^6.1.6",
     "@react-spring/web": "^9.7.3",
     "@redis/client": "^1.2.0",
-    "@sentry/nextjs": "^8.5.0",
+    "@sentry/nextjs": "^8.36.0",
     "@use-gesture/react": "^10.3.0",
     "autosuggest-highlight": "^3.3.4",
     "classnames": "^2.3.1",
@@ -41,13 +43,13 @@
     "image-size": "^1.0.1",
     "isomorphic-dompurify": "^0.18.0",
     "mongodb": "^5.5.0",
-    "next": "^14.2.9",
+    "next": "15.0.2",
     "node-fetch": "^3.2.4",
     "node-vibrant": "^3.2.1-alpha.1",
     "plaiceholder": "^2.3.0",
-    "react": "^18.3.1",
+    "react": "18.3.1",
     "react-device-detect": "^2.2.2",
-    "react-dom": "^18.3.1",
+    "react-dom": "18.3.1",
     "sharp": "^0.33.3",
     "smoothscroll-polyfill": "^0.4.4",
     "swr": "^2.2.4",
@@ -59,15 +61,15 @@
     "@testing-library/react": "^16.0.0",
     "@types/autosuggest-highlight": "^3.2.0",
     "@types/elasticsearch": "^5.0.40",
-    "@types/node": "17.0.32",
     "@types/jest": "^29.5.12",
-    "@types/react": "^18.3.3",
+    "@types/node": "17.0.32",
+    "@types/react": "18.3.12",
     "@types/smoothscroll-polyfill": "^0.3.1",
     "@types/url-file-size": "^1.0.0",
     "@typescript-eslint/eslint-plugin": "^5.42.1",
     "eslint": "8.15.0",
     "eslint-config-airbnb": "^19.0.4",
-    "eslint-config-next": "12.1.6",
+    "eslint-config-next": "15.0.2",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-import": "^2.27.5",
     "eslint-plugin-jsx-a11y": "^6.5.1",
@@ -95,8 +97,11 @@
   },
   "browserslist": {
     "production": [
-      "chrome 49",
-      "opera 36"
+      "chrome 109",
+      "opera 95"
     ]
+  },
+  "overrides": {
+    "@types/react": "18.3.12"
   }
 }
